refactor: extract KnowledgeIngestionService from KnowledgeStoreService
- Create knowledge-ingestion-service.ts with all ingestion methods
- Move rebuildIndex, background embedding/HyPE, compaction, and reflection/output ingestion
- KnowledgeStoreService delegates to ingestion service (public API unchanged)
- knowledge-store-service.ts shrinks by 413 lines (exceeds 250-line target)
- Add dedicated ingest route handler
